Move ErtScript workflow
