+++ src/ert/config/workflow_job.py
@@ -2,6 +2,8 @@ from __future__ import annotations
 
 import logging
 import os
+from argparse import ArgumentParser
+from collections.abc import Callable
 from dataclasses import dataclass
 from typing import TypeAlias
 
@@ -131,3 +133,80 @@ class WorkflowJob:
             raise ValueError(f"Unknown job type {c} in {self}")
 
         return list(map(content_to_type, self.arg_types))
+
+
+class ErtScriptWorkflow(WorkflowJob):
+    """
+    Single workflow configuration object
+    """
+
+    def __init__(
+        self, ertscript_class: type[ErtScript], name: str | None = None
+    ) -> None:
+        """
+        :param ertscript_class: Class inheriting from ErtScript
+        :param name: Optional name for workflow, default is class name
+        """
+        self.source_package = self._get_source_package(ertscript_class)
+        self._description = ertscript_class.__doc__ or ""
+        self._examples: str | None = None
+        self._parser: Callable[[], ArgumentParser] | None = None
+        self._category = "other"
+        super().__init__(
+            name=self._get_func_name(ertscript_class, name),
+            ert_script=ertscript_class,
+            min_args=None,
+            max_args=None,
+            arg_types=[],
+            executable=None,
+        )
+
+    @property
+    def description(self) -> str:
+        """
+        A string of valid rst, will be added to the documentation
+        """
+        return self._description
+
+    @description.setter
+    def description(self, description: str) -> None:
+        self._description = description
+
+    @property
+    def examples(self) -> str | None:
+        """
+        A string of valid rst, will be added to the documentation
+        """
+        return self._examples
+
+    @examples.setter
+    def examples(self, examples: str | None) -> None:
+        self._examples = examples
+
+    @property
+    def parser(self) -> Callable[[], ArgumentParser] | None:
+        return self._parser
+
+    @parser.setter
+    def parser(self, parser: Callable[[], ArgumentParser] | None) -> None:
+        self._parser = parser
+
+    @property
+    def category(self) -> str:
+        """
+        A dot separated string
+        """
+        return self._category
+
+    @category.setter
+    def category(self, category: str) -> None:
+        self._category = category
+
+    @staticmethod
+    def _get_func_name(func: type[ErtScript], name: str | None) -> str:
+        return name or func.__name__
+
+    @staticmethod
+    def _get_source_package(module: type[ErtScript]) -> str:
+        base, _, _ = module.__module__.partition(".")
+        return base
