+++ tests/ert/unit_tests/plugins/test_workflow_config.py
@@ -4,12 +4,13 @@ from unittest.mock import Mock
 import pytest
 
 import ert
+import ert.config.workflow_job
 from ert.plugins import workflow_config
 
 
 def test_workflow_config_duplicate_log_message(caplog, monkeypatch):
     def get_mock_config():
-        workflow_mock = Mock(spec=workflow_config.ErtScriptWorkflow)
+        workflow_mock = Mock(spec=ert.config.workflow_job.ErtScriptWorkflow)
         workflow_mock.name = "same_name"
         return workflow_mock
 
@@ -32,7 +33,7 @@ def test_workflow_config_init_name(monkeypatch, name, expected):
     mock_func.__name__ = "default_name"
 
     monkeypatch.setattr(ert.config.workflow_job.WorkflowJob, "__post_init__", Mock())
-    workflow = workflow_config.ErtScriptWorkflow(mock_func, name=name)
+    workflow = ert.config.workflow_job.ErtScriptWorkflow(mock_func, name=name)
 
     assert workflow.name == expected
     assert workflow.ert_script == mock_func
